Moved around files
Most notably, put search and its related stuff (not search results) into src/routes/home as those are the homepage
