+++ src/styles/components/media/Poster.tsx
