+++ src/styles/reuse/generalStylesInteractions.tsx
